refactor(abw): decompose blurb composer
